feat(alerts): Convert alert list to functional, useApiQuery
