+++ static/app/views/alerts/list/rules/alertRulesList.spec.tsx
@@ -1,49 +1,25 @@
-import type {Location} from 'history';
-
 import {initializeOrg} from 'sentry-test/initializeOrg';
 import {act, render, screen, userEvent, within} from 'sentry-test/reactTestingLibrary';
 
 import OrganizationStore from 'sentry/stores/organizationStore';
 import ProjectsStore from 'sentry/stores/projectsStore';
 import TeamStore from 'sentry/stores/teamStore';
-import {Organization} from 'sentry/types';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import AlertRulesList from 'sentry/views/alerts/list/rules';
 import {IncidentStatus} from 'sentry/views/alerts/types';
-import {OrganizationContext} from 'sentry/views/organizationContext';
+
+import AlertRulesList from './alertRulesList';
 
 jest.mock('sentry/utils/analytics');
 
 describe('AlertRulesList', () => {
-  const {routerContext, organization, router} = initializeOrg({
-    organization: {
-      access: ['alerts:write'],
-    },
+  const defaultOrg = TestStubs.Organization({
+    access: ['alerts:write'],
   });
   TeamStore.loadInitialData([TestStubs.Team()], false, null);
-  let rulesMock;
-  let projectMock;
+  let rulesMock!: jest.Mock;
+  let projectMock!: jest.Mock;
   const pageLinks =
     '<https://sentry.io/api/0/organizations/org-slug/combined-rules/?cursor=0:0:1>; rel="previous"; results="false"; cursor="0:0:1", ' +
     '<https://sentry.io/api/0/organizations/org-slug/combined-rules/?cursor=0:100:0>; rel="next"; results="true"; cursor="0:100:0"';
-
-  const getComponent = (
-    props: {location?: Location; organization?: Organization} = {}
-  ) => (
-    <OrganizationContext.Provider value={props.organization ?? organization}>
-      <AlertRulesList
-        {...TestStubs.routeComponentProps()}
-        organization={props.organization ?? organization}
-        params={{}}
-        location={TestStubs.location({query: {}, search: ''})}
-        router={router}
-        {...props}
-      />
-    </OrganizationContext.Provider>
-  );
-
-  const createWrapper = (props = {}) =>
-    render(getComponent(props), {context: routerContext});
 
   beforeEach(() => {
     rulesMock = MockApiClient.addMockResponse({
@@ -81,7 +57,7 @@ describe('AlertRulesList', () => {
       ],
     });
 
-    act(() => OrganizationStore.onUpdate(organization, {replace: true}));
+    act(() => OrganizationStore.onUpdate(defaultOrg, {replace: true}));
     act(() => ProjectsStore.loadInitialData([]));
   });
 
@@ -92,7 +68,8 @@ describe('AlertRulesList', () => {
   });
 
   it('displays list', async () => {
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
 
@@ -104,13 +81,6 @@ describe('AlertRulesList', () => {
     );
 
     expect(screen.getAllByTestId('badge-display-name')[0]).toHaveTextContent('earth');
-
-    expect(trackAnalytics).toHaveBeenCalledWith(
-      'alert_rules.viewed',
-      expect.objectContaining({
-        sort: 'incident_status,date_triggered',
-      })
-    );
   });
 
   it('displays empty state', async () => {
@@ -118,8 +88,8 @@ describe('AlertRulesList', () => {
       url: '/organizations/org-slug/combined-rules/',
       body: [],
     });
-
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     expect(
       await screen.findByText('No alert rules found for the current query.')
@@ -129,7 +99,8 @@ describe('AlertRulesList', () => {
   });
 
   it('displays team dropdown context if unassigned', async () => {
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     const assignee = (await screen.findAllByTestId('alert-row-assignee'))[0];
     const btn = within(assignee).getAllByRole('button')[0];
 
@@ -148,7 +119,9 @@ describe('AlertRulesList', () => {
       url: '/projects/org-slug/earth/rules/123/',
       body: [],
     });
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
+
     const assignee = (await screen.findAllByTestId('alert-row-assignee'))[0];
     const btn = within(assignee).getAllByRole('button')[0];
 
@@ -167,7 +140,8 @@ describe('AlertRulesList', () => {
   });
 
   it('displays dropdown context menu with actions', async () => {
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     const actions = (await screen.findAllByRole('button', {name: 'Actions'}))[0];
     expect(actions).toBeInTheDocument();
 
@@ -179,7 +153,10 @@ describe('AlertRulesList', () => {
   });
 
   it('sends user to new alert page on duplicate action', async () => {
-    createWrapper();
+    const {routerContext, organization, router} = initializeOrg({
+      organization: defaultOrg,
+    });
+    render(<AlertRulesList />, {context: routerContext, organization});
     const actions = (await screen.findAllByRole('button', {name: 'Actions'}))[0];
     expect(actions).toBeInTheDocument();
 
@@ -202,28 +179,24 @@ describe('AlertRulesList', () => {
   });
 
   it('sorts by name', async () => {
-    const {rerender} = createWrapper();
-
-    // The name column is not used for sorting
-    expect(await screen.findByText('Alert Rule')).toHaveAttribute('aria-sort', 'none');
-
-    // Sort by the name column
-    rerender(
-      getComponent({
+    const {routerContext, organization} = initializeOrg({
+      organization: defaultOrg,
+      router: {
         location: TestStubs.location({
           query: {asc: '1', sort: 'name'},
+          // Sort by the name column
           search: '?asc=1&sort=name`',
         }),
-      })
-    );
+      },
+    });
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     expect(await screen.findByText('Alert Rule')).toHaveAttribute(
       'aria-sort',
       'ascending'
     );
 
-    expect(rulesMock).toHaveBeenCalledTimes(2);
-
+    expect(rulesMock).toHaveBeenCalledTimes(1);
     expect(rulesMock).toHaveBeenCalledWith(
       '/organizations/org-slug/combined-rules/',
       expect.objectContaining({
@@ -234,20 +207,18 @@ describe('AlertRulesList', () => {
 
   it('disables the new alert button for members', async () => {
     const noAccessOrg = {
-      ...organization,
+      ...defaultOrg,
       access: [],
     };
-
-    render(getComponent({organization: noAccessOrg}), {
-      context: TestStubs.routerContext([{organization: noAccessOrg}]),
-      organization: noAccessOrg,
-    });
+    const {routerContext, organization} = initializeOrg({organization: noAccessOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     expect(await screen.findByLabelText('Create Alert')).toBeDisabled();
   });
 
   it('searches by name', async () => {
-    createWrapper();
+    const {routerContext, organization, router} = initializeOrg();
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     const search = await screen.findByPlaceholderText('Search by name');
     expect(search).toBeInTheDocument();
@@ -259,27 +230,23 @@ describe('AlertRulesList', () => {
       expect.objectContaining({
         query: {
           name: testQuery,
-          expand: ['latestIncident', 'lastTriggered'],
-          sort: ['incident_status', 'date_triggered'],
-          team: ['myteams', 'unassigned'],
         },
       })
     );
   });
 
   it('uses empty team query parameter when removing all teams', async () => {
-    const {rerender} = createWrapper();
-
-    expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
-
-    rerender(
-      getComponent({
+    const {routerContext, organization, router} = initializeOrg({
+      router: {
         location: TestStubs.location({
           query: {team: 'myteams'},
           search: '?team=myteams`',
         }),
-      })
-    );
+      },
+    });
+    render(<AlertRulesList />, {context: routerContext, organization});
+
+    expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
 
     await userEvent.click(await screen.findByRole('button', {name: 'My Teams'}));
 
@@ -290,8 +257,6 @@ describe('AlertRulesList', () => {
     expect(router.push).toHaveBeenCalledWith(
       expect.objectContaining({
         query: {
-          expand: ['latestIncident', 'lastTriggered'],
-          sort: ['incident_status', 'date_triggered'],
           team: '',
         },
       })
@@ -299,7 +264,8 @@ describe('AlertRulesList', () => {
   });
 
   it('displays metric alert status', async () => {
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     const rules = await screen.findAllByText('My Incident Rule');
 
     expect(rules[0]).toBeInTheDocument();
@@ -322,7 +288,8 @@ describe('AlertRulesList', () => {
         }),
       ],
     });
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
     expect(screen.getByText('Disabled')).toBeInTheDocument();
   });
@@ -341,7 +308,8 @@ describe('AlertRulesList', () => {
         }),
       ],
     });
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
     expect(screen.getByText('Disabled')).toBeInTheDocument();
     expect(screen.queryByText('Muted')).not.toBeInTheDocument();
@@ -359,7 +327,8 @@ describe('AlertRulesList', () => {
         }),
       ],
     });
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
     expect(screen.getByText('Muted')).toBeInTheDocument();
   });
@@ -375,13 +344,15 @@ describe('AlertRulesList', () => {
         }),
       ],
     });
-    createWrapper();
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
     expect(await screen.findByText('My Incident Rule')).toBeInTheDocument();
     expect(screen.getByText('Muted')).toBeInTheDocument();
   });
 
   it('sorts by alert rule', async () => {
-    createWrapper({organization});
+    const {routerContext, organization} = initializeOrg({organization: defaultOrg});
+    render(<AlertRulesList />, {context: routerContext, organization});
 
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
 
@@ -398,10 +369,10 @@ describe('AlertRulesList', () => {
   });
 
   it('preserves empty team query parameter on pagination', async () => {
-    createWrapper({
-      organization,
-      location: {query: {team: ''}, search: '?team=`'},
+    const {routerContext, organization, router} = initializeOrg({
+      organization: defaultOrg,
     });
+    render(<AlertRulesList />, {context: routerContext, organization});
     expect(await screen.findByText('First Issue Alert')).toBeInTheDocument();
 
     await userEvent.click(screen.getByLabelText('Next'));
@@ -409,8 +380,6 @@ describe('AlertRulesList', () => {
     expect(router.push).toHaveBeenCalledWith(
       expect.objectContaining({
         query: {
-          expand: ['latestIncident', 'lastTriggered'],
-          sort: ['incident_status', 'date_triggered'],
           team: '',
           cursor: '0:100:0',
         },
