+++ static/app/views/alerts/list/rules/row.tsx
@@ -443,10 +443,10 @@ const IssueAlertStatusWrapper = styled('div')`
 `;
 
 const AlertNameWrapper = styled('div')<{isIssueAlert?: boolean}>`
+  ${p => p.theme.overflowEllipsis}
   display: flex;
   align-items: center;
   gap: ${space(2)};
-  position: relative;
   ${p => p.isIssueAlert && `padding: ${space(3)} ${space(2)}; line-height: 2.4;`}
 `;
 
@@ -458,16 +458,6 @@ const AlertNameAndStatus = styled('div')`
 const AlertName = styled('div')`
   ${p => p.theme.overflowEllipsis}
   font-size: ${p => p.theme.fontSizeLarge};
-
-  @media (max-width: ${p => p.theme.breakpoints.xlarge}) {
-    max-width: 300px;
-  }
-  @media (max-width: ${p => p.theme.breakpoints.large}) {
-    max-width: 165px;
-  }
-  @media (max-width: ${p => p.theme.breakpoints.medium}) {
-    max-width: 100px;
-  }
 `;
 
 const AlertIncidentDate = styled('div')`
